move op_fetch_signed_headers to common, add wasm route
