+++ crates/dcl_deno/src/js/fetch/mod.rs
@@ -3,9 +3,9 @@ use std::{cell::RefCell, rc::Rc};
 mod fetch_response_body_resource;
 
 use bevy::{asset::AsyncReadExt, prelude::debug};
-use common::{rpc::RpcCall, structs::SceneMeta};
+use common::rpc::RpcCall;
 use deno_core::{
-    anyhow::{self, anyhow},
+    anyhow,
     error::{type_error, AnyError},
     futures::{FutureExt, TryStreamExt},
     op2, AsyncRefCell, BufView, ByteString, CancelHandle, JsBuffer, OpDecl, OpState, Resource,
@@ -16,18 +16,15 @@ use deno_net::NetPermissions;
 use deno_web::TimersPermission;
 use http::{
     header::{ACCEPT_ENCODING, CONTENT_LENGTH, HOST, RANGE},
-    HeaderName, HeaderValue, Method, Uri,
+    HeaderName, HeaderValue, Method,
 };
 use ipfs::IpfsResource;
 use serde::{Deserialize, Serialize};
 
 use fetch_response_body_resource::FetchResponseBodyResource;
 use tokio::sync::oneshot::channel;
-use wallet::{sign_request, Wallet};
 
 use dcl::{interface::crdt_context::CrdtContext, RpcCalls};
-
-use dcl::js::runtime::realm_information;
 
 // we have to provide fetch perm structs even though we don't use them
 pub struct FP;
@@ -350,27 +347,6 @@ pub fn op_fetch_custom_client(
     Ok(state.resource_table.add(ClientResource(builder.build()?)))
 }
 
-#[derive(Serialize, Default, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct SignedFetchMetaRealm {
-    hostname: String,
-    protocol: String,
-    server_name: String,
-}
-
-#[derive(Serialize, Default, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct SignedFetchMeta {
-    origin: Option<String>,
-    scene_id: Option<String>,
-    parcel: Option<String>,
-    tld: Option<String>,
-    network: Option<String>,
-    is_guest: Option<bool>,
-    realm: SignedFetchMetaRealm,
-    signer: String,
-}
-
 #[op2(async)]
 #[serde]
 pub async fn op_signed_fetch_headers(
@@ -378,51 +354,7 @@ pub async fn op_signed_fetch_headers(
     #[string] uri: String,
     #[string] method: Option<String>,
 ) -> Result<Vec<(String, String)>, AnyError> {
-    debug!("op_signed_fetch_headers");
-
-    let is_preview = state.borrow().borrow::<CrdtContext>().preview;
-    let scheme = Uri::try_from(&uri)?;
-    let scheme = scheme.scheme_str();
-    if !is_preview && !([Some("https"), Some("wss")].contains(&scheme)) {
-        anyhow::bail!("URL scheme must be `https` (request `{}`)", uri);
-    }
-
-    let realm_info = realm_information(state.clone()).await?;
-    let wallet = state.borrow().borrow::<Wallet>().clone();
-    let urn = state.borrow().borrow::<CrdtContext>().hash.clone();
-    let ipfs = state.borrow().borrow::<IpfsResource>().clone();
-    let scene_meta = ipfs
-        .entity_definition(&urn)
-        .await
-        .and_then(|(entity, _)| {
-            serde_json::from_str::<SceneMeta>(&entity.metadata.unwrap_or_default()).ok()
-        })
-        .ok_or(anyhow!("failed to parse scene metadata"))?;
-
-    let meta = SignedFetchMeta {
-        origin: Some(realm_info.base_url.clone()),
-        scene_id: Some(urn),
-        parcel: Some(scene_meta.scene.base.clone()),
-        tld: Some("org".to_owned()),
-        network: Some("mainnet".to_owned()),
-        is_guest: Some(wallet.is_guest()),
-        realm: SignedFetchMetaRealm {
-            hostname: realm_info.base_url,
-            protocol: "v3".to_owned(),
-            server_name: realm_info.realm_name,
-        },
-        signer: "decentraland-kernel-scene".to_owned(),
-    };
-
-    debug!("signed fetch meta {:?}", meta);
-
-    sign_request(
-        method.as_deref().unwrap_or("get"),
-        &Uri::try_from(uri)?,
-        &wallet,
-        meta,
-    )
-    .await
+    dcl::js::fetch::op_signed_fetch_headers(state, uri, method).await
 }
 
 use core::future::Future;
